// ares/src/app/shared/components/division/models/division-model.ts
export interface DivisionModel {
    id: string;
    name: string;
    nickname: string;
    registry: string;
    email: string;
    representative_id: string;
    partner_id: string;
    address: string;
    phone: string;
    city: string;
    theme: string;
    expiresDate: Date;
    lockedDate?: any;
    createdby: string;
    updatedby: string;
    created: Date;
    updated: Date;
}